+++ src/evaluator.rs
@@ -1,21 +1,25 @@
+use base64::{CharacterSet, Config};
 use charset::Charset;
+use std::{result, string};
+use thiserror::Error;
 
-use crate::parser::{Encoding, ParsedEncodedWord, ParsedEncodedWords, ClearText};
+use crate::parser::{ClearText, Encoding, ParsedEncodedWord, ParsedEncodedWords};
 
-pub type Result<T> = std::result::Result<T, Error>;
-
-#[derive(thiserror::Error, Debug)]
+#[derive(Error, Debug)]
 pub enum Error {
     #[error(transparent)]
-    DecodeUtf8(#[from] std::string::FromUtf8Error),
+    DecodeUtf8Error(#[from] string::FromUtf8Error),
     #[error(transparent)]
-    DecodeBase64(#[from] base64::DecodeError),
+    DecodeBase64Error(#[from] base64::DecodeError),
     #[error(transparent)]
-    DecodeQuotedPrintable(#[from] quoted_printable::QuotedPrintableError),
+    DecodeQuotedPrintableError(#[from] quoted_printable::QuotedPrintableError),
 }
 
+type Result<T> = result::Result<T, Error>;
+
 fn decode_base64(encoded_bytes: Vec<u8>) -> Result<Vec<u8>> {
-    let decoded_bytes = base64::decode(encoded_bytes)?;
+    let config = Config::new(CharacterSet::Standard, true).decode_allow_trailing_bits(true);
+    let decoded_bytes = base64::decode_config(&encoded_bytes, config)?;
     Ok(decoded_bytes)
 }
 
@@ -39,12 +43,10 @@ fn decode_with_encoding(encoding: Encoding, encoded_bytes: Vec<u8>) -> Result<Ve
     match encoding {
         Encoding::B => decode_base64(encoded_bytes),
         Encoding::Q => decode_quoted_printable(encoded_bytes),
-
     }
 }
 
 fn decode_with_charset(charset: Option<Charset>, decoded_bytes: Vec<u8>) -> Result<String> {
-
     let decoded_str = match charset {
         Some(charset) => charset.decode(&decoded_bytes).0,
         None => charset::decode_ascii(&decoded_bytes),
@@ -71,16 +73,14 @@ fn decode_parsed_encoded_word(
 pub fn run(parsed_encoded_words: ParsedEncodedWords) -> Result<String> {
     let message = parsed_encoded_words
         .into_iter()
-        .flat_map(
-            |parsed_encoded_word: ParsedEncodedWord| match parsed_encoded_word {
-                ParsedEncodedWord::ClearText(clear_text) => decode_utf8_string(clear_text),
-                ParsedEncodedWord::EncodedWord {
-                    charset,
-                    encoding,
-                    encoded_text,
-                } => decode_parsed_encoded_word(charset, encoding, encoded_text),
-            },
-        )
+        .flat_map(|parsed_encoded_word| match parsed_encoded_word {
+            ParsedEncodedWord::ClearText(clear_text) => decode_utf8_string(clear_text),
+            ParsedEncodedWord::EncodedWord {
+                charset,
+                encoding,
+                encoded_text,
+            } => decode_parsed_encoded_word(charset, encoding, encoded_text),
+        })
         .collect();
 
     Ok(message)